models/segmenting/requirements.txt
beautifulsoup4==4.9.3
cycler==0.10.0
kiwisolver==1.3.1
lxml==4.6.3
matplotlib==3.4.2
numpy==1.21.0
pandas==1.3.0
Pillow==8.3.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
soupsieve==2.2.1
torch==1.8.0+cu111
torchaudio==0.8.0
torchvision==0.9.0+cu111
typing-extensions==3.10.0.0
